+++ trixnity-client/trixnity-client-media-indexeddb/src/jsMain/kotlin/net/folivo/trixnity/client/media/indexeddb/IndexedDBMediaStore.kt
@@ -3,11 +3,17 @@ package net.folivo.trixnity.client.media.indexeddb
 import com.juul.indexeddb.Database
 import com.juul.indexeddb.Key
 import com.juul.indexeddb.openDatabase
-import js.typedarrays.toUint8Array
+import js.typedarrays.Uint8Array
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.launch
 import net.folivo.trixnity.client.media.MediaStore
 import net.folivo.trixnity.utils.ByteArrayFlow
-import net.folivo.trixnity.utils.toByteArray
-import net.folivo.trixnity.utils.toByteArrayFlow
+import net.folivo.trixnity.utils.byteArrayFlowFromReadableStream
+import net.folivo.trixnity.utils.writeTo
+import web.blob.Blob
+import web.http.BodyInit
+import web.http.Response
+import web.streams.TransformStream
 
 class IndexedDBMediaStore(val databaseName: String = "trixnity_media") : MediaStore {
     companion object {
@@ -25,8 +31,12 @@ class IndexedDBMediaStore(val databaseName: String = "trixnity_media") : MediaSt
         }
     }
 
-    override suspend fun addMedia(url: String, content: ByteArrayFlow) {
-        val value = content.toByteArray().toUint8Array()
+    override suspend fun addMedia(url: String, content: ByteArrayFlow) = coroutineScope {
+        val transformStream = TransformStream<Uint8Array, Uint8Array>()
+        launch {
+            content.writeTo(transformStream.writable)
+        }
+        val value = Response(BodyInit(transformStream.readable)).blob()
         database.writeTransaction(MEDIA_OBJECT_STORE_NAME) {
             val store = objectStore(MEDIA_OBJECT_STORE_NAME)
             store.put(value, Key(url))
@@ -37,7 +47,8 @@ class IndexedDBMediaStore(val databaseName: String = "trixnity_media") : MediaSt
         database.transaction(MEDIA_OBJECT_STORE_NAME) {
             val store = objectStore(MEDIA_OBJECT_STORE_NAME)
             store.get(Key(url))
-        }.unsafeCast<ByteArray?>()?.toByteArrayFlow()
+        }.unsafeCast<Blob?>()
+            ?.let { byteArrayFlowFromReadableStream { it.stream() } }
 
     override suspend fun deleteMedia(url: String): Unit =
         database.writeTransaction(MEDIA_OBJECT_STORE_NAME) {